libft fini

+++ ft_memmove.c
@@ -14,7 +14,7 @@
 
 void	*ft_memmove(void *dest, const void *src, size_t n)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	if (n == 0 || (dest == NULL && src == NULL))
@@ -22,7 +22,7 @@ void	*ft_memmove(void *dest, const void *src, size_t n)
 	if (dest > src)
 	{
 		i = n - 1;
-		while (i >= 0)
+		while ((int)i >= 0)
 		{
 			*(char *)(dest + i) = *(char *)(src + i);
 			i--;
@@ -30,7 +30,7 @@ void	*ft_memmove(void *dest, const void *src, size_t n)
 	}
 	else
 	{
-		while (i < (int)n)
+		while (i < n)
 		{
 			*(char *)(dest + i) = *(char *)(src + i);
 			i++;
